requirements.txt
pyodbc==5.0.1
pandas==2.1.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
schedule==1.2.0

